requirements.txt
ipykernel
numpy
pandas
matplotlib
seaborn
statsmodels
scipy
scikit-learn
python-dotenv
pymongo
setuptools
certifi
pymongo[srv]==3.12

# -e .